refactor(tests): move catch2 u8string to_string func
